requirements.txt
faiss_cpu==1.9.0.post1
fastapi==0.115.8
Flask==3.1.0
flask_sqlalchemy==3.1.1
google_api_python_client==2.159.0
joblib==1.4.2
mysql_connector_repackaged==0.3.1
numpy==2.2.2
Office365_REST_Python_Client==2.5.14
ortools==9.11.4210
pandas==2.2.3
pdfkit==1.0.0
plotly==5.24.1
protobuf==5.29.3
pydantic==2.10.6
PyPDF2==3.0.1
python_docx==1.1.2
pdfkit==1.0.0

# Utilidades
python-dotenv==1.0.1
Requests==2.32.3
schedule==1.2.2
tqdm==4.67.1
joblib==1.4.2
pydantic==2.10.6
setuptools==65.5.0

# Integración con servicios
google_api_python_client==2.159.0
Office365_REST_Python_Client==2.5.14

# Optimización
ortools==9.11.4210

# Testing
pytest==7.4.3
pytest-cov==4.1.0
